Test : BuyAmountTest 생성
